convert expectations!'s integrations tests to mock! and automock! tests
